modified:   .gitignore modified:   .vscode/launch.json new file:   Untitled.ipynb new file:   Untitled1.ipynb renamed:    templates/main.html -> app/templates/main.html renamed:    templates/sucess.html -> app/templates/sucess.html renamed:    templates/upload.html -> app/templates/upload.html renamed:    templates/uploadmessage.html -> app/templates/uploadmessage.html renamed:    db_utilites/mongodb.py -> app/utilities/db_utilities/mongodb.py renamed:    faiss_db.py -> app/utilities/faiss_db.py renamed:    logger.py -> app/utilities/logger.py new file:   app/utilities/processpdf.py renamed:    vectorstore/index.faiss -> app/vectorstore/index.faiss renamed:    vectorstore/index.pkl -> app/vectorstore/index.pkl modified:   main.py new file:   temp/deeplearning-ian_goodfellow.pdf modified:   test.py
