+++ .gitignore
@@ -166,7 +166,7 @@ dmypy.json
 cython_debug/
 
 documents
-
+embedding_model/
 
 # PyCharm
 #  JetBrains specific template is maintained in a separate JetBrains.gitignore that can
