+++ .vscode/launch.json
@@ -11,7 +11,6 @@
             "module": "uvicorn",
             "args": [
                 "main:app",
-                "--reload",
                 "--port",
                 "4023",
                 "--host",
