--- templates/main.html
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>PDF Database</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 0;
-            display: flex;
-            flex-direction: column;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            background-color: #f4f4f4;
-        }
-        .header {
-            text-align: center;
-            margin-bottom: 20px;
-        }
-        .header h1 {
-            margin: 0;
-            font-size: 2.5em;
-        }
-        .header p {
-            margin: 0;
-            font-size: 1.2em;
-            color: #666;
-        }
-        .form-container {
-            background-color: #fff;
-            padding: 20px;
-            border-radius: 8px;
-            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-        }
-        .form-group {
-            display: flex;
-            justify-content: space-between;
-            margin-bottom: 15px;
-        }
-        .form-group button {
-            padding: 10px 20px;
-            font-size: 16px;
-            background-color: #007BFF;
-            color: white;
-            border: none;
-            border-radius: 4px;
-            cursor: pointer;
-            flex: 1;
-            margin: 0 5px;
-        }
-        .form-group button:hover {
-            background-color: #0056b3;
-        }
-    </style>
-</head>
-<body>
-    <div class="header">
-        <h1>PDF Database</h1>
-        <p>Manage and Query Your PDF Documents Efficiently</p>
-    </div>
-    <div class="form-container">
-        <form action="/upload" method="get">
-            <div class="form-group">
-                <button type="submit">Upload</button>
-            </div>
-        </form>
-        <form action="/query" method="POST">
-            <div class="form-group">
-                <button type="submit">Query</button>
-            </div>
-        </form>
-    </div>
-</body>
-</html>
