--- templates/sucess.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Success Page</title>
-    <style>
-        body {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            margin: 0;
-            font-family: Arial, sans-serif;
-            background-color: #f4f4f9;
-        }
-        .message-box {
-            padding: 20px;
-            border: 1px solid #28a745;
-            background-color: #d4edda;
-            color: #155724;
-            border-radius: 5px;
-            text-align: center;
-        }
-    </style>
-</head>
-<body>
-    <div class="message-box">
-        <h1>Success!</h1>
-        <p>Your operation was successful.</p>
-    </div>
-</body>
-</html>
