--- templates/upload.html
@@ -1,67 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document Upload</title>
-        <style>
-            body {
-                font-family: Arial, sans-serif;
-                background-color: #f0f0f0;
-                padding: 20px;
-            }
-            h1 {
-                text-align: center;
-                color: #333;
-            }
-            form {
-                max-width: 400px;
-                margin: 0 auto;
-                background-color: #fff;
-                padding: 20px;
-                border-radius: 8px;
-                box-shadow: 0 0 10px rgba(0,0,0,0.1);
-            }
-            label {
-                display: block;
-                margin-bottom: 10px;
-                color: #555;
-            }
-            input[type="text"], input[type="file"] {
-                padding: 10px;
-                margin-bottom: 10px;
-                border: 1px solid #ccc;
-                border-radius: 4px;
-                width: 100%;
-            }
-            button[type="submit"] {
-                background-color: #007bff;
-                color: #fff;
-                border: none;
-                padding: 10px 20px;
-                cursor: pointer;
-                border-radius: 4px;
-                width: 100%;
-            }
-            button[type="submit"]:hover {
-                background-color: #0056b3;
-            }
-        </style>
-    </head>
-    <body>
-        <h1>Upload a Document</h1>
-        <form action="/processfile/" method="post" enctype="multipart/form-data">
-            <label for="file_id">File ID:</label>
-            <input type="text" id="file_id" name="file_id" required>
-    
-            <label for="file_topic">File Topic:</label>
-            <input type="text" id="file_topic" name="file_topic" required>
-    
-            <label for="file">Choose a document to upload:</label>
-            <input type="file" id="file" name="file" accept=".doc,.docx,.pdf" required>
-            
-            <button type="submit">Upload</button>
-        </form>
-    </body>
-    </html>
